requirements.txt
streamlit
google-generativeai
langchain
langchain-google-genai 
python-dotenv
